Fixed Halvers
Removed some duplicate / unneccesary files regarding smartboxes, and fixed the halver smartbox.
