+++ interface/smartbox/halver.config
@@ -1,0 +1,32 @@
+{
+	"gui" : {
+		"background" : {
+			"type" : "background",
+			"fileHeader" : "/interface/smartbox/header.png",
+			"fileBody" : "/interface/smartbox/halver.png"
+		},
+		"close" : {
+			"type" : "button",
+			"base" : "/interface/x.png",
+			"hover" : "/interface/xhover.png",
+			"press" : "/interface/xpress.png",
+			//dx: 18, dy: 24
+			"position" : [111, 56]
+		},
+		"itemGrid" : {
+			"type" : "itemgrid",
+			"backingImage" : "/interface/inventory/empty.png",
+			"position" : [10, 17],
+			"dimensions" : [1, 1],
+			"spacing" : [22, 20]
+		},
+		"outputItemGrid" : {
+			"type" : "itemgrid",
+			"backingImage" : "/interface/inventory/empty.png",
+			"position" : [54, 17],
+			"dimensions" : [2, 1],
+			"spacing" : [44, 20],
+			"slotOffset" : 1
+		}
+	}
+}
