+++ interface/smartbox/singular.config
@@ -2,8 +2,8 @@
 	"gui" : {
 		"background" : {
 			"type" : "background",
-			"fileHeader" : "/interface/smartbox/singular/header.png",
-			"fileBody" : "/interface/smartbox/singular/body.png"
+			"fileHeader" : "/interface/smartbox/header.png",
+			"fileBody" : "/interface/smartbox/singular.png"
 		},
 		"close" : {
 			"type" : "button",
