+++ interface/smartbox/splitter.config
@@ -2,8 +2,8 @@
 	"gui" : {
 		"background" : {
 			"type" : "background",
-			"fileHeader" : "/interface/smartbox/splitter/header.png",
-			"fileBody" : "/interface/smartbox/splitter/body.png"
+			"fileHeader" : "/interface/smartbox/header.png",
+			"fileBody" : "/interface/smartbox/splitter.png"
 		},
 		"close" : {
 			"type" : "button",
