+++ interface/smartbox/splitter.png
