--- interface/smartbox/splitter/splitter.lua
@@ -1,11 +1,0 @@
-function init()
-
-end
-
-function update()
-
-end
-
-function refresh()
-
-end
