+++ objects/smartbox/halver/halver.object
@@ -10,7 +10,7 @@
 	
 	"description" : "Splits a stack into 50% of its previous value",
 	"shortdescription" : "Stack Halver",
-	"subtitle" : "Subtitle Test",
+	"subtitle" : "Smartbox",
 	
 	"inventoryIcon" : "icon.png",
 	"orientations" : [
@@ -32,11 +32,11 @@
 	"acceptsItems" : true,
 	
 	"openSounds" : [ "/sfx/objects/locker_open.ogg" ],
-	"slotCount" : 4,
-	"uiConfig" : "/interface/smartbox/splitter/splitter.config",
+	"slotCount" : 3,
+	"uiConfig" : "/interface/smartbox/halver.config",
 	"frameCooldown" : 67,
 	"autoCloseCooldown" : 3600,
 	
-	"scripts" : [ "/objects/smartbox/splitter/halver.lua" ],
+	"scripts" : [ "/objects/smartbox/halver/halver.lua" ],
 	"scriptDelta" : 1
 }
