+++ objects/smartbox/singular/singular.object
@@ -11,7 +11,7 @@
 	"description" : "Storage with only one slot!
 Item-Network compatible.",
 	"shortdescription" : "Single-Slot Storage",
-	"subtitle" : "Subtitle Test",
+	"subtitle" : "Smartbox",
 	
 	"inventoryIcon" : "icon.png",
 	"orientations" : [
@@ -34,7 +34,7 @@ Item-Network compatible.",
 	
 	"openSounds" : [ "/sfx/objects/locker_open.ogg" ],
 	"slotCount" : 1,
-	"uiConfig" : "/interface/smartbox/singular/singular.config",
+	"uiConfig" : "/interface/smartbox/singular.config",
 	"frameCooldown" : 67,
 	"autoCloseCooldown" : 3600,
