+++ objects/smartbox/splitter/splitter.object
@@ -12,7 +12,7 @@
 and 1s (as well as it can).
 Item-Network compatible.",
 	"shortdescription" : "Stack Splitter",
-	"subtitle" : "Subtitle Test",
+	"subtitle" : "Smartbox",
 	
 	"inventoryIcon" : "icon.png",
 	"orientations" : [
@@ -35,7 +35,7 @@ Item-Network compatible.",
 	
 	"openSounds" : [ "/sfx/objects/locker_open.ogg" ],
 	"slotCount" : 4,
-	"uiConfig" : "/interface/smartbox/splitter/splitter.config",
+	"uiConfig" : "/interface/smartbox/splitter.config",
 	"frameCooldown" : 67,
 	"autoCloseCooldown" : 3600,
